api_server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
websockets==12.0
matlabengine==24.1.2
numpy==1.26.3
